requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
defusedxml==0.7.1
dotenv==0.9.9
et_xmlfile==2.0.0
idna==3.10
jira==3.10.5
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
